Now able to configure cross compiler version

+++ RaspiQtCrossCompile_BuildGccV2.sh
@@ -9,6 +9,41 @@ NC='\033[0m' # No Color
 BUILD_LOC=$1
 SOURCE_CACHE_LOC=$2
 
+target="aarch64-linux-gnu"
+# target="arm-linux-gnueabihf"
+target_os="bookworm"
+# target_os="buster"
+
+# This is the configuration for the Bookworm version
+gcc_ver="12.2.0"
+binutils_ver="2.40"
+glibc_ver="2.36"
+
+# Check for overrides
+if [ $# -ge 6 ]; then
+    gcc_ver=$3
+    binutils_ver=$4
+    glibc_ver=$5
+    target=$6
+    target_os=$7
+fi
+
+base_install_loc="/opt/cross-pi-gcc/$target_os"
+
+echo -e "${GREEN}Configuration:${NC}"
+echo -e "${BLUE}GCC Ver: ${gcc_ver}${NC}"
+echo -e "${BLUE}Binutils Ver: ${binutils_ver}${NC}"
+echo -e "${BLUE}GCC Ver: ${glibc_ver}${NC}"
+echo -e "${BLUE}Target arch: ${target}${NC}"
+echo -e "${BLUE}Target OS: ${target_os}${NC}"
+echo -e "${BLUE}Install loc: ${base_install_loc}/$target${NC}"
+
+read -p "Is this correct(y/n)" input
+if [[ $input != [Yy]* ]]; then
+    echo -e "${RED}Aborting${NC}"
+    exit 1
+fi
+
 # Extra, but not necessary
 sudo apt-get -y install help2man gettext
 
@@ -23,29 +58,29 @@ download_sources () {
 
     cd ${SOURCE_CACHE_LOC}
     # Download gcc sources
-    if [ ! -f binutils-2.40.tar.bz2 ]; then
+    if [ ! -f binutils-$binutils_ver.tar.bz2 ]; then
         echo -e "${BLUE}Downloading binutils source${NC}"
-        wget https://ftpmirror.gnu.org/binutils/binutils-2.40.tar.bz2
+        wget https://ftpmirror.gnu.org/binutils/binutils-$binutils_ver.tar.bz2
 
         # Handle error
         if [ $? -ne 0 ]; then
-            echo -e "${RED}Failed to download bin utils source${NC}"
+            echo -e "${RED}Failed to download bin utils source: $binutils_ver${NC}"
             return 1
         fi
     fi
-    if [ ! -f glibc-2.36.tar.bz2 ]; then
+    if [ ! -f glibc-$glibc_ver.tar.bz2 ]; then
         echo -e "${BLUE}Downloading glibc source${NC}"
-        wget https://ftpmirror.gnu.org/glibc/glibc-2.36.tar.bz2
+        wget https://ftpmirror.gnu.org/glibc/glibc-$glibc_ver.tar.bz2
 
         # Handle error
         if [ $? -ne 0 ]; then
-            echo -e "${RED}Failed to download glibc source${NC}"
+            echo -e "${RED}Failed to download glibc source: $glibc_ver${NC}"
             return 1
         fi
     fi
-    if [ ! -f gcc-12.2.0.tar.gz ]; then
+    if [ ! -f gcc-$gcc_ver.tar.gz ]; then
     echo -e "${BLUE}Downloading gcc source${NC}"
-        wget https://ftpmirror.gnu.org/gcc/gcc-12.2.0/gcc-12.2.0.tar.gz
+        wget https://ftpmirror.gnu.org/gcc/gcc-$gcc_ver/gcc-$gcc_ver.tar.gz
 
         # Handle error
         if [ $? -ne 0 ]; then
@@ -84,10 +119,10 @@ extract_sources () {
         fi
 
         # Remove files from install loc files from 
-        if [ -d /opt/cross-pi-gcc ]; then
+        if [ -d $base_install_loc ]; then
             echo -e "${BLUE}Removing old gcc dir${NC}"
             
-            sudo rm -rf /opt/cross-pi-gcc
+            sudo rm -rf $base_install_loc
 
             # Handle error if applicable
             if [ $? -ne 0 ]; then
@@ -101,7 +136,7 @@ extract_sources () {
     mkdir -p gcc_all && cd gcc_all
 
     echo -e "${BLUE}Starting extract binutils${NC}"
-    tar xf ${SOURCE_CACHE_LOC}/binutils-2.40.tar.bz2
+    tar xf ${SOURCE_CACHE_LOC}/binutils-${binutils_ver}.tar.bz2
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to extract bin utils source${NC}"
@@ -109,7 +144,7 @@ extract_sources () {
     fi
 
     echo -e "${BLUE}Starting extract glibc${NC}"
-    tar xf ${SOURCE_CACHE_LOC}/glibc-2.36.tar.bz2
+    tar xf ${SOURCE_CACHE_LOC}/glibc-${glibc_ver}.tar.bz2
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to extract glibc source${NC}"
@@ -117,7 +152,7 @@ extract_sources () {
     fi
 
     echo -e "${BLUE}Starting extract gcc${NC}"
-    tar xf ${SOURCE_CACHE_LOC}/gcc-12.2.0.tar.gz
+    tar xf ${SOURCE_CACHE_LOC}/gcc-${gcc_ver}.tar.gz
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to extract gcc source${NC}"
@@ -134,19 +169,19 @@ extract_sources () {
 create_install_dir () {
     echo -e "${GREEN}Creating install dir${NC}"
 
-    echo -e "${BLUE}sudo mkdir -p /opt/cross-pi-gcc${NC}"
-    sudo mkdir -p /opt/cross-pi-gcc
+    echo -e "${BLUE}sudo mkdir -p $base_install_loc${NC}"
+    sudo mkdir -p $base_install_loc
 
     if [ $? -ne 0 ]; then 
         echo -e "${RED}Failed to make install dir${NC}"
         return 1
     fi
 
-    echo -e "${BLUE}sudo chown $USER /opt/cross-pi-gcc${NC}"
-    sudo chown $USER /opt/cross-pi-gcc
+    echo -e "${BLUE}sudo chown $USER $base_install_loc${NC}"
+    sudo chown $USER $base_install_loc
 
     echo -e "${BLUE}Adding to path${NC}"
-    export PATH=/opt/cross-pi-gcc/bin:$PATH
+    export PATH=$base_install_loc/bin:$PATH
 
     echo -e "${GREEN}Finished creating install dir${NC}"
 
@@ -156,7 +191,7 @@ create_install_dir () {
 download_gcc_prereqs () {
     echo -e "${GREEN}Starting download prereqs${NC}"
 
-    cd ${BUILD_LOC}/gcc_all/gcc-12.2.0
+    cd ${BUILD_LOC}/gcc_all/gcc-$1
 
     contrib/download_prerequisites
 
@@ -180,8 +215,8 @@ install_kernel_headers () {
     echo -e "${BLUE}Setting kernel to ${KERNEL}${NC}"
     export KERNEL=kernel_2712
 
-    echo -e "${BLUE}Starting headers_install: make ARCH=arm64 INSTALL_HDR_PATH=/opt/cross-pi-gcc/aarch64-linux-gnu headers_install${NC}"
-    make ARCH=arm64 INSTALL_HDR_PATH=/opt/cross-pi-gcc/aarch64-linux-gnu headers_install
+    echo -e "${BLUE}Starting headers_install: make ARCH=arm64 INSTALL_HDR_PATH=$base_install_loc/$target headers_install${NC}"
+    make ARCH=arm64 INSTALL_HDR_PATH=$base_install_loc/$target headers_install
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to install rpi kernel headers${NC}"
@@ -203,8 +238,8 @@ build_binutils_step_1 () {
     mkdir build-binutils && cd build-binutils
 
     echo -e "${BLUE}Configuring bin utils${NC}"
-    echo -e "${BLUE}../binutils-2.40/configure --prefix=/opt/cross-pi-gcc --target=aarch64-linux-gnu --with-arch=armv8 --disable-multilib${NC}"
-    ../binutils-2.40/configure --prefix=/opt/cross-pi-gcc --target=aarch64-linux-gnu --with-arch=armv8 --disable-multilib
+    echo -e "${BLUE}$BUILD_LOC/gcc_all/binutils-$1/configure --prefix=$base_install_loc --target=$target --with-arch=armv8 --disable-multilib${NC}"
+    $BUILD_LOC/gcc_all/binutils-$1/configure --prefix=$base_install_loc --target=$target --with-arch=armv8 --disable-multilib
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to configure bin utils${NC}"
@@ -213,7 +248,7 @@ build_binutils_step_1 () {
 
     echo -e "${BLUE}Starting build bin utils${NC}"
     echo -e "${BLUE}make -j$threads${NC}"
-    make -j$threads
+    make -j$threads -s
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to make binutils${NC}"
@@ -237,14 +272,30 @@ build_binutils_step_1 () {
 patch_asan_linux () {
     echo -e "${GREEN}Starting patch asan_linux${NC}"
 
-    if ! grep -q "#define PATH_MAX" ../gcc-12.2.0/libsanitizer/asan/asan_linux.cpp; then
-        echo -e "${BLUE}Patching asan_linux${NC}"
-        sed -i.back '67i #ifndef PATH_MAX' ../gcc-12.2.0/libsanitizer/asan/asan_linux.cpp
-        sed -i.back '68i #define PATH_MAX 4096' ../gcc-12.2.0/libsanitizer/asan/asan_linux.cpp
-        sed -i.back '69i #endif' ../gcc-12.2.0/libsanitizer/asan/asan_linux.cpp
-    else
-        echo -e "${BLUE}The file has already been patched${NC}"
+    if [ -f "$BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cpp" ]; then
+        if ! grep -q "#define PATH_MAX" $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cpp; then
+            echo -e "${BLUE}Patching asan_linux${NC}"
+            sed -i.back '67i #ifndef PATH_MAX' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cpp
+            sed -i.back '68i #define PATH_MAX 4096' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cpp
+            sed -i.back '69i #endif' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cpp
+        else
+            echo -e "${BLUE}The file has already been patched${NC}"
+        fi
+    elif [ -f "$BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cc" ]; then
+        if ! grep -q "#define PATH_MAX" $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cc; then
+            echo -e "${BLUE}Patching asan_linux${NC}"
+            sed -i.back '67i #ifndef PATH_MAX' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cc
+            sed -i.back '68i #define PATH_MAX 4096' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cc
+            sed -i.back '69i #endif' $BUILD_LOC/gcc_all/gcc-$1/libsanitizer/asan/asan_linux.cc
+        else
+            echo -e "${BLUE}The file has already been patched${NC}"
+        fi
     fi
+
+    if [ $? -ne 0 ]; then
+        echo -e "${RED}Failed to patch asan${NC}"
+        return 1
+    fi 
 
     echo -e "${GREEN}Finished patch asan_linux${NC}"
     return 0
@@ -260,8 +311,8 @@ build_gcc_step_2 () {
     mkdir build-gcc && cd build-gcc
 
     echo -e "${BLUE}Configuring gcc${NC}"
-    echo -e "${BLUE}../gcc-12.2.0/configure --prefix=/opt/cross-pi-gcc --target=aarch64-linux-gnu --enable-languages=c,c++ --disable-multilib${NC}"
-    ../gcc-12.2.0/configure --prefix=/opt/cross-pi-gcc --target=aarch64-linux-gnu --enable-languages=c,c++ --disable-multilib
+    echo -e "${BLUE}$BUILD_LOC/gcc_all/gcc-$1/configure --prefix=$base_install_loc --target=$target --enable-languages=c,c++ --disable-multilib${NC}"
+    $BUILD_LOC/gcc_all/gcc-$1/configure --prefix=$base_install_loc --target=$target --enable-languages=c,c++ --disable-multilib
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to configure gcc${NC}"
@@ -270,7 +321,7 @@ build_gcc_step_2 () {
 
     echo -e "${BLUE}Starting build gcc${NC}"
     echo -e "${BLUE}make -j $threads all-gcc${NC}"
-    make -j$threads all-gcc
+    make -s -j$threads all-gcc
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to make gcc${NC}"
@@ -301,8 +352,8 @@ continue_gcc_step_3 () {
     mkdir build-glibc && cd build-glibc
 
     echo -e "${BLUE}Configuring glibc${NC}"
-    echo -e "${BLUE}../glibc-2.36/configure --prefix=/opt/cross-pi-gcc/aarch64-linux-gnu --build=$MACHTYPE --host=aarch64-linux-gnu --target=aarch64-linux-gnu --with-headers=/opt/cross-pi-gcc/aarch64-linux-gnu/include --disable-multilib libc_cv_forced_unwind=yes${NC}"
-    ../glibc-2.36/configure --prefix=/opt/cross-pi-gcc/aarch64-linux-gnu --build=$MACHTYPE --host=aarch64-linux-gnu --target=aarch64-linux-gnu --with-headers=/opt/cross-pi-gcc/aarch64-linux-gnu/include --disable-multilib libc_cv_forced_unwind=yes
+    echo -e "${BLUE}$BUILD_LOC/gcc_all/glibc-$1/configure --prefix=$base_install_loc/$target-linux-gnu --build=$MACHTYPE --host=aarch64-linux-gnu --target=$target --with-headers=$base_install_loc/$target/include --disable-multilib libc_cv_forced_unwind=yes${NC}"
+    $BUILD_LOC/gcc_all/glibc-$1/configure --prefix=$base_install_loc/$target --build=$MACHTYPE --host=aarch64-linux-gnu --target=$target --with-headers=$base_install_loc/$target/include --disable-multilib libc_cv_forced_unwind=yes
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to configure glibc${NC}"
@@ -313,32 +364,37 @@ continue_gcc_step_3 () {
     echo -e "${BLUE}make install-bootstrap-headers=yes install-headers${NC}"
     make install-bootstrap-headers=yes install-headers
 
+    if [ $? -ne 0 ]; then
+        echo -e "${RED}Failed on: make install-bootstrap-headers=yes install-headers${NC}"
+        return 1
+    fi
+
     echo -e "${BLUE}make -j$threads csu/subdir_lib${NC}"
-    make -j$threads csu/subdir_lib
+    make -j$threads csu/subdir_lib -Wno-error
 
     if [ $? -ne 0 ]; then
-        echo -e "${RED}Failed on: install csu/crt1.o csu/crti.o csu/crtn.o /opt/cross-pi-gcc/aarch64-linux-gnu/lib${NC}"
+        echo -e "${RED}Failed on: make -j$threads csu/subdir_lib${NC}"
         return 1
     fi
 
-    echo -e "${BLUE}install csu/crt1.o csu/crti.o csu/crtn.o /opt/cross-pi-gcc/aarch64-linux-gnu/lib${NC}"
-    install csu/crt1.o csu/crti.o csu/crtn.o /opt/cross-pi-gcc/aarch64-linux-gnu/lib
+    echo -e "${BLUE}install csu/crt1.o csu/crti.o csu/crtn.o $base_install_loc/$target/lib${NC}"
+    install csu/crt1.o csu/crti.o csu/crtn.o $base_install_loc/$target/lib
 
     if [ $? -ne 0 ]; then
-        echo -e "${RED}Failed on: install csu/crt1.o csu/crti.o csu/crtn.o /opt/cross-pi-gcc/aarch64-linux-gnu/lib${NC}"
+        echo -e "${RED}Failed on: install csu/crt1.o csu/crti.o csu/crtn.o $base_install_loc/$target/lib${NC}"
         return 1
     fi
 
-    echo -e "${BLUE}aarch64-linux-gnu-gcc -nostdlib -nostartfiles -shared -x c /dev/null -o /opt/cross-pi-gcc/aarch64-linux-gnu/lib/libc.so${NC}"
-    aarch64-linux-gnu-gcc -nostdlib -nostartfiles -shared -x c /dev/null -o /opt/cross-pi-gcc/aarch64-linux-gnu/lib/libc.so
+    echo -e "${BLUE}$target-gcc -nostdlib -nostartfiles -shared -x c /dev/null -o $base_install_loc/$target/lib/libc.so${NC}"
+    $target-gcc -nostdlib -nostartfiles -shared -x c /dev/null -o $base_install_loc/$target/lib/libc.so
 
     if [ $? -ne 0 ]; then
-        echo -e "${RED}Failed on: install csu/crt1.o csu/crti.o csu/crtn.o /opt/cross-pi-gcc/aarch64-linux-gnu/lib${NC}"
+        echo -e "${RED}Failed on: install csu/crt1.o csu/crti.o csu/crtn.o $base_install_loc/$target/lib${NC}"
         return 1
     fi
 
-    echo -e "${BLUE}touch /opt/cross-pi-gcc/aarch64-linux-gnu/include/gnu/stubs.h${NC}"
-    touch /opt/cross-pi-gcc/aarch64-linux-gnu/include/gnu/stubs.h
+    echo -e "${BLUE}touch $base_install_loc/$target/include/gnu/stubs.h${NC}"
+    touch $base_install_loc/$target/include/gnu/stubs.h
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to create stubs.h file${NC}"
@@ -356,7 +412,7 @@ continue_build_gcc_step_4 () {
     cd ${BUILD_LOC}/gcc_all/build-gcc
 
     echo -e "${BLUE}make -j${threads} -s all-target-libgcc${NC}"
-    make -j${threads} -s all-target-libgcc
+    make -s -j${threads} -s all-target-libgcc
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed on: make -j${threads} -s all-target-libgcc${NC}"
@@ -383,7 +439,7 @@ finish_build_glibc_step_5 () {
 
     echo -e "${BLUE}Building glibc${NC}"
     echo -e "${BLUE}make -j${threads}${NC}"
-    make -j${threads}
+    make -s -j${threads}
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to build glibc${NC}"
@@ -411,7 +467,7 @@ finish_build_gcc_step_6 () {
 
     echo -e "${BLUE}Making gcc${NC}"
     echo -e "${BLUE}make -j${threads}${NC}"
-    make -j${threads}
+    make -j${threads} -s
 
     if [ $? -ne 0 ]; then
         echo -e "${RED}Failed to make gcc${NC}"
@@ -453,7 +509,7 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-download_gcc_prereqs
+download_gcc_prereqs ${gcc_ver}
 
 if [ $? -ne 0 ]; then
     echo -e "${RED}Failed to build gcc${NC}"
@@ -467,28 +523,28 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-build_binutils_step_1
+build_binutils_step_1 $binutils_ver
 
 if [ $? -ne 0 ]; then
     echo -e "${RED}Failed to build gcc${NC}"
     exit 1
 fi
 
-patch_asan_linux
+patch_asan_linux $gcc_ver
 
 if [ $? -ne 0 ]; then
     echo -e "${RED}Failed to build gcc${NC}"
     exit 1
 fi
 
-build_gcc_step_2
+build_gcc_step_2 $gcc_ver
 
 if [ $? -ne 0 ]; then
     echo -e "${RED}Failed to build gcc${NC}"
     exit 1
 fi
 
-continue_gcc_step_3
+continue_gcc_step_3 $glibc_ver
 
 if [ $? -ne 0 ]; then
     echo -e "${RED}Failed to build gcc${NC}"
@@ -518,4 +574,4 @@ fi
 
 # Test the cross compiler
 echo -e "${GREEN}Testing the cross compiler${NC}"
-aarch64-linux-gnu-gcc --version
+$target-gcc --version
